reorg RethEnv methods
