+++ crates/tn-reth/src/lib.rs
@@ -138,11 +138,10 @@ pub use reth_transaction_pool::{
 mod cli;
 pub mod dirs;
 pub mod payload;
-use payload::TNPayload;
 pub mod traits;
 pub mod txn_pool;
 pub use txn_pool::*;
-use worker::WorkerNetwork;
+mod env;
 pub mod error;
 mod evm;
 pub mod forward;
@@ -155,6 +154,7 @@ pub mod worker;
 pub use cli::{
     init_txpool_defaults, RethCommand, RethConfig, TN_TXPOOL_MAX_ACCOUNT_SLOTS_PER_SENDER,
 };
+pub use env::*;
 #[cfg(feature = "faucet")]
 pub use evm::faucet_mint_role_slot;
 #[cfg(not(feature = "faucet"))]
@@ -165,6 +165,7 @@ pub use evm::{
     BLS_G1_PRECOMPILE_ADDRESS, TELCOIN_PRECOMPILE_ADDRESS,
 };
 pub use forward::WorkerRpcForwarder;
+pub use metrics::report_db_metrics;
 pub use types::*;
 
 #[cfg(any(feature = "test-utils", test))]
@@ -178,72 +179,6 @@ static BASEFEE_ADDRESS: OnceLock<Address> = OnceLock::new();
 /// Note the basefee address is set once for the chain and will not change (outside of a hard fork).
 pub fn basefee_address() -> Address {
     *BASEFEE_ADDRESS.get().unwrap_or(&GOVERNANCE_SAFE_ADDRESS)
-}
-
-/// Set the basefee address.  This will only work on the first call and should be during program
-/// initialization. Calling more than once will do nothing, not calling early can lead to an unset
-/// basefee address and a chain fork.
-fn set_basefee_address(address: Option<Address>) {
-    // Ignore the error. Should probably panic on error but this will break some test environments.
-    let _ = BASEFEE_ADDRESS.set(address.unwrap_or(GOVERNANCE_SAFE_ADDRESS));
-}
-
-/// One transaction in the chain-wide sequential transaction feed.
-///
-/// Storage assigns every mined transaction a sequential [`TxNumber`]; this entry pairs the
-/// recovered transaction with that number and the containing block's context so consumers can
-/// serve "latest N transactions" style queries without touching block bodies of empty blocks.
-#[derive(Clone, Debug)]
-pub struct TxFeedEntry {
-    /// Sequential, chain-wide transaction number (position in the global feed).
-    pub tx_number: TxNumber,
-    /// The signed transaction with its sender attached.
-    pub transaction: Recovered<TransactionSigned>,
-    /// Number of the block containing this transaction.
-    pub block_number: BlockNumber,
-    /// Hash of the block containing this transaction.
-    pub block_hash: B256,
-    /// Timestamp of the containing block.
-    pub timestamp: u64,
-    /// Zero-based index of the transaction within its block.
-    pub index: u64,
-}
-
-/// This is a wrapped abstraction around Reth.
-///
-/// It should allow the telcoin app to access the required functionality without
-/// leaking Reth internals all over the codebase (this makes staying up to date
-/// VERY time consuming).
-///
-/// `RethEnv` wraps its fields in an `Arc<RethEnvInner>` so that cloning is cheap
-/// (just an `Arc` bump) and the type is trivially `Send + Sync`.
-#[derive(Clone)]
-pub struct RethEnv {
-    /// The inner state wrapped in an Arc for cheap cloning and thread safety.
-    inner: Arc<RethEnvInner>,
-}
-
-/// Inner state for [`RethEnv`].
-///
-/// This struct holds all the actual fields and is wrapped in an `Arc` by `RethEnv`.
-struct RethEnvInner {
-    /// The type that holds all information needed to launch the node's engine.
-    ///
-    /// The [NodeConfig] is reth-specific and holds many helper functions that
-    /// help TN stay in-sync with the Ethereum community.
-    node_config: NodeConfig<RethChainSpec>,
-    /// Type that fetches data from the database.
-    blockchain_provider: BlockchainProvider<TelcoinNode>,
-    /// The type to configure the EVM for execution.
-    evm_config: TnEvmConfig,
-    /// The type to spawn tasks.
-    task_spawner: TaskSpawner,
-}
-
-impl std::fmt::Debug for RethEnv {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "RethEnv, config: {:?}", self.inner.node_config)
-    }
 }
 
 /// Wrapper for Reth ChainSpec, just a layer of abstraction.
@@ -284,1875 +219,13 @@ impl ChainSpec {
     }
 }
 
-/// Report sampled reth database metrics (table sizes, page usage, freelist).
-///
-/// Intended as a pre-scrape hook for the prometheus metrics endpoint. Lives here to keep
-/// reth-db types out of the node crate.
-pub fn report_db_metrics(db: &RethDb) {
-    use reth_db::database_metrics::DatabaseMetrics;
-    db.report_metrics();
-}
-
-impl RethEnv {
-    /// Create a new Reth DB.
-    /// Break this out so this can be created upfront and used even on a
-    /// restart (when catching up for instance).
-    pub fn new_database<P: AsRef<Path>>(
-        reth_config: &RethConfig,
-        db_path: P,
-    ) -> eyre::Result<RethDb> {
-        let db_path = db_path.as_ref();
-        info!(target: "tn::reth", path = ?db_path, "opening database");
-        // with_metrics: record per-operation db latency metrics (noop unless the global
-        // metrics recorder is installed, i.e. the node runs with `--metrics`)
-        Ok(Arc::new(init_db(db_path, reth_config.0.db.database_args())?.with_metrics()))
-    }
-
-    /// Produce a new wrapped Reth environment from a config, DB path and task manager.
-    ///
-    /// This method MUST be called from within a tokio runtime.
-    pub fn new(
-        reth_config: &RethConfig,
-        task_manager: &TaskManager,
-        database: RethDb,
-        basefee_address: Option<Address>,
-        rewards_counter: RewardsCounter,
-    ) -> eyre::Result<Self> {
-        let node_config = reth_config.0.clone();
-        let evm_config = TnEvmConfig::new(reth_config.0.chain.clone(), rewards_counter);
-        let provider_factory = Self::init_provider_factory(&node_config, database)?;
-        let blockchain_provider = BlockchainProvider::new(provider_factory.clone())?;
-        let task_spawner = task_manager.get_spawner();
-        set_basefee_address(basefee_address);
-        // baseline the block-building counters at zero now, while the recorder is known to be
-        // installed, so a node that never drops a transaction still exports both series
-        crate::metrics::init();
-
-        Ok(Self {
-            inner: Arc::new(RethEnvInner {
-                node_config,
-                blockchain_provider,
-                evm_config,
-                task_spawner,
-            }),
-        })
-    }
-
-    /// Initialize the provider factory and related components
-    fn init_provider_factory(
-        node_config: &NodeConfig<RethChainSpec>,
-        database: Arc<DatabaseEnv>,
-    ) -> eyre::Result<ProviderFactory<TelcoinNode>> {
-        // Initialize provider factory with static files
-        let datadir = node_config.datadir();
-        let rocksdb_provider = reth_provider::providers::RocksDBProvider::new(datadir.data_dir())?;
-        let runtime = reth_tasks::Runtime::with_existing_handle(tokio::runtime::Handle::current())?;
-        let provider_factory = ProviderFactory::new(
-            database,
-            Arc::clone(&node_config.chain),
-            StaticFileProvider::read_write(datadir.static_files())?,
-            rocksdb_provider,
-            runtime,
-        )?;
-
-        // Initialize genesis if needed
-        let genesis_hash = init_genesis(&provider_factory)?;
-        debug!(target: "tn::execution", chain=%node_config.chain.chain, ?genesis_hash, "Initialized genesis");
-
-        Ok(provider_factory)
-    }
-
-    /// Initialize a new transaction pool for worker.
-    pub fn init_txn_pool(&self) -> eyre::Result<WorkerTxPool> {
-        WorkerTxPool::new(
-            &self.inner.node_config,
-            &self.inner.task_spawner,
-            &self.inner.blockchain_provider,
-            &self.inner.evm_config,
-        )
-    }
-
-    /// Return a channel reciever that will return each canonical block in turn.
-    pub fn canonical_block_stream(&self) -> CanonStateNotificationStream {
-        self.inner.blockchain_provider.canonical_state_stream()
-    }
-
-    /// Return a reference to the [TaskSpawner] for spawning tasks.
-    pub fn get_task_spawner(&self) -> &TaskSpawner {
-        &self.inner.task_spawner
-    }
-
-    /// Return the chainspec for this instance.
-    pub fn chainspec(&self) -> ChainSpec {
-        ChainSpec(self.inner.node_config.chain.clone())
-    }
-
-    /// Return the canonical in-memory state.
-    pub fn canonical_in_memory_state(&self) -> CanonicalInMemoryState {
-        self.inner.blockchain_provider.canonical_in_memory_state()
-    }
-
-    /// Construct a canonical block from a worker's block that reached consensus.
-    pub fn build_block_from_batch_payload(
-        &self,
-        payload: TNPayload,
-        transactions: &Vec<Vec<u8>>,
-        anchor_hash: B256,
-        ancestors: &[DeferredTrieData],
-    ) -> TnRethResult<ExecutedBlock> {
-        let parent_header = payload.parent_header.clone();
-        debug!(target: "engine", ?parent_header, "retrieving state for next block");
-        let state_provider =
-            self.inner.blockchain_provider.state_by_block_hash(parent_header.hash())?;
-        let state = StateProviderDatabase::new(&state_provider);
-        let mut cached_reads = CachedReads::default();
-        let mut db = State::builder()
-            .with_database(cached_reads.as_db_mut(state))
-            .with_bundle_update()
-            .build();
-
-        debug!(
-            target: "engine",
-            parent = ?parent_header.num_hash(),
-            "building new payload"
-        );
-
-        // copy in case of error
-        let batch_digest = payload.batch_digest;
-
-        let mut builder =
-            self.inner.evm_config.builder_for_next_block(&mut db, &parent_header, payload)?;
-
-        builder.apply_pre_execution_changes().inspect_err(|err| {
-            warn!(target: "engine", %err, "failed to apply pre-execution changes");
-        })?;
-
-        // Phase 1: Recover all transactions (ECDSA ecrecover) in parallel via rayon.
-        // Always use par_iter — the slight overhead on small batches is negligible
-        // compared to the savings on large ones, and avoids an extra code path.
-        //
-        // A transaction whose signer cannot be recovered cannot be executed, but a
-        // certified sub-DAG is fixed and identical on every honest node, so returning
-        // an error here would deterministically halt (and, on restart-replay,
-        // crash-loop) the whole network on a single undecodable transaction. Instead
-        // drop the unrecoverable transactions and continue, mirroring the `InvalidTx`
-        // tolerance of the execute phase below: every node drops the same
-        // transactions from the same certified bytes, so the resulting block stays
-        // deterministic (issue #933). The primary defense is validating batches before
-        // they can be certified; this only bounds the blast radius if one ever is.
-        let recovered_txs = transactions
-            .par_iter()
-            .filter_map(|tx_bytes| {
-                reth_recover_raw_transaction::<TransactionSigned>(tx_bytes)
-                    .inspect_err(|e| {
-                        error!(
-                            target: "engine",
-                            batch=?batch_digest,
-                            ?tx_bytes,
-                            "failed to recover signer, dropping transaction: {e}"
-                        )
-                    })
-                    .ok()
-            })
-            .collect::<Vec<_>>();
-
-        // Recovery failures are the signal that a batch carrying undecodable transaction bytes
-        // was certified, so make them alertable instead of log-only. Count once, after the
-        // parallel phase: incrementing inside the rayon closure would contend a single atomic on
-        // every transaction, and the batch width that sets the cost of that contention is
-        // attacker-chosen. The difference is exact because `filter_map` is the only thing above
-        // that removes elements. Unconditional, including the overwhelmingly common zero: one
-        // relaxed atomic add per block is free next to executing the block, and a guard here
-        // would leave the series unregistered on a healthy node.
-        let unrecoverable = transactions.len().saturating_sub(recovered_txs.len());
-        RETH_METRICS
-            .unrecoverable_txs_dropped_total
-            .increment(u64::try_from(unrecoverable).unwrap_or(u64::MAX));
-
-        // Phase 2: Execute recovered transactions sequentially.
-        for recovered in recovered_txs {
-            // forks are impossible
-            match builder.execute_transaction(recovered.clone()) {
-                Ok(_gas_used) => (),
-                Err(BlockExecutionError::Validation(BlockValidationError::InvalidTx {
-                    error,
-                    ..
-                })) => {
-                    // allow transaction errors (ie - duplicates)
-                    //
-                    // it's possible that another worker's batch included this transaction
-                    warn!(target: "engine", %error,  "skipping invalid transaction: {:#?}", recovered);
-                    // expected in normal operation - see the field docs, this is not an alert
-                    RETH_METRICS.invalid_txs_skipped_total.increment(1);
-                    continue;
-                }
-                // this is an error that we should treat as fatal for this attempt
-                Err(err) => return Err(err.into()),
-            }
-        }
-
-        let BlockBuilderOutcome { execution_result, block, hashed_state, trie_updates } =
-            builder.finish(&state_provider)?;
-
-        debug!(target: "engine", hash=?block.hash(), "block builder outcome");
-        let block_execution_output =
-            BlockExecutionOutput { result: execution_result, state: db.take_bundle() };
-        let computed_trie_data = DeferredTrieData::sort_and_build_trie_input(
-            Arc::new(hashed_state),
-            Arc::new(trie_updates),
-            anchor_hash,
-            ancestors,
-        );
-        let res: ExecutedBlock<TNPrimitives> = ExecutedBlock::new(
-            Arc::new(block),
-            Arc::new(block_execution_output),
-            computed_trie_data,
-        );
-
-        Ok(res)
-    }
-
-    /// Finalize the block (header) executed from consensus output in memory.
-    ///
-    /// The finalized/safe database markers are persisted atomically with the blocks in
-    /// [`Self::finish_executing_output`]; this method only updates the in-memory
-    /// finalized/safe watches (consumed by components like RPC) and prunes persisted blocks
-    /// from the canonical in-memory state.
-    pub fn finalize_block(&self, header: SealedHeader) -> TnRethResult<()> {
-        let num_hash = header.num_hash();
-        // this clears up old blocks in-memory
-        self.inner.blockchain_provider.set_finalized(header.clone());
-
-        // update safe block last because this is less time sensitive but still needs to happen
-        self.inner.blockchain_provider.set_safe(header);
-
-        // cleanup chain state in memory
-        // this returns the `canonical_chain().count()` back to 0
-        self.inner
-            .blockchain_provider
-            .canonical_in_memory_state()
-            .remove_persisted_blocks(num_hash);
-
-        Ok(())
-    }
-
-    /// Advance a lagging finalized marker to the persisted canonical tip.
-    ///
-    /// Blocks and the finalized/safe markers commit in a single database transaction in
-    /// [`Self::finish_executing_output`], so a crash can no longer leave the marker behind the
-    /// persisted tip — the two-transaction crash window this heal was written for does not
-    /// exist for new writes. The heal is retained as defense-in-depth for databases written by
-    /// pre-fix versions, which committed the markers separately in `finalize_block`: a crash
-    /// between the two transactions restarted the node with `finalized marker < persisted
-    /// canonical tip`, hiding the gap blocks from every marker reader (accumulator catchup,
-    /// the tx-pool's last-seen seed, the RPC `finalized`/`safe` tags).
-    ///
-    /// Advancing the marker is sound because every persisted canonical block is consensus-final
-    /// by construction: blocks only enter the canonical database from committed consensus
-    /// output, so there is no speculative or reorgable segment the marker could overrun. The
-    /// lag is purely a persistence artifact of the pre-fix two-transaction design.
-    ///
-    /// The heal persists the marker itself ([`Self::finalize_block`] no longer writes the
-    /// database) and then routes through `finalize_block` for the in-memory half: the
-    /// finalized/safe watches are seeded from the (stale) database at provider construction,
-    /// and startup marker readers like the accumulator catchup consult the watch — a DB-only
-    /// write would leave them reading the stale value.
-    ///
-    /// Replay is unaffected: the missed-consensus watermark derives from the persisted
-    /// execution tip (which this method never moves), not the finalized marker, so healing
-    /// cannot cause `replay_missed_consensus` to re-forward committed output — no
-    /// double-execution is possible.
-    ///
-    /// Call once at startup, before anything reads the marker and before any consensus output
-    /// executes. A marker ahead of the tip fails with
-    /// [`TnRethError::FinalizedMarkerAheadOfTip`]: the execution database lost blocks this node
-    /// already attested as final.
-    ///
-    /// Like every startup tip reader, this trusts the persisted tip itself; a torn write
-    /// between MDBX and static files below the provider is a pre-existing exposure this method
-    /// neither adds to nor guards against.
-    pub fn heal_finalized_to_persisted_tip(&self) -> TnRethResult<()> {
-        // One RO provider = one consistent MDBX snapshot for all three reads. Read errors
-        // PROPAGATE: the public last_block_number()/last_finalized_block_number() wrappers
-        // unwrap_or(0) and would misdiagnose a failed read as tip=0 here.
-        let provider = self.inner.blockchain_provider.database_provider_ro()?;
-        let tip = provider.last_block_number()?;
-        // None = never finalized (fresh genesis). Some(0) is unreachable in production (the
-        // first marker write covers block >= 1), so collapsing None to 0 is safe.
-        let finalized = provider.last_finalized_block_number()?.unwrap_or(0);
-
-        if finalized > tip {
-            return Err(TnRethError::FinalizedMarkerAheadOfTip { finalized, tip });
-        }
-        if finalized == tip {
-            // Normal restart (marker caught up), or fresh genesis (marker never written and
-            // tip = 0): genesis intentionally stays unfinalized so
-            // finalized_block_hash_number_for_startup keeps its genesis fallback.
-            return Ok(());
-        }
-
-        let header =
-            provider.sealed_header(tip)?.ok_or(ProviderError::HeaderNotFound(tip.into()))?;
-        drop(provider);
-        warn!(
-            target: "tn::reth",
-            finalized,
-            tip,
-            "finalized marker lags the persisted canonical tip (database written by a pre-fix \
-             version that committed blocks and the marker separately); healing marker to tip"
-        );
-        // durably fix the marker on disk before touching the watches: finalize_block only
-        // updates in-memory state, so a lagging pre-fix database needs its rows rewritten here
-        let provider_rw = self.inner.blockchain_provider.database_provider_rw()?;
-        provider_rw.save_finalized_block_number(tip)?;
-        provider_rw.save_safe_block_number(tip)?;
-        provider_rw.commit()?;
-        self.finalize_block(header)
-    }
-
-    /// This makes all blocks canonical, commits them to the database,
-    /// broadcasts new chain on `canon_state_notification_sender`
-    /// and set last executed header as the tracked header.
-    ///
-    /// The finalized/safe markers commit in the same database transaction as the blocks, so a
-    /// crash can never leave the persisted marker behind the persisted tip. The in-memory
-    /// finalized/safe watches are updated afterwards by [`Self::finalize_block`].
-    ///
-    /// It also clears the canonical in-memory state.
-    pub fn finish_executing_output(
-        &self,
-        blocks: Vec<ExecutedBlock>,
-        engine_update: Option<(Round, ConsensusNumHash, tokio::sync::mpsc::Sender<EngineUpdate>)>,
-    ) -> TnRethResult<()> {
-        // NOTE: this makes all blocks canonical, commits them to the database,
-        // and broadcasts new chain on `canon_state_notification_sender`
-        //
-        // the canon_state_notifications include every block executed in this round
-        //
-        // the worker's pool maintenance task subcribes to these events
-        debug!(
-            target: "engine",
-            first=?blocks.first().map(|b| b.recovered_block.num_hash()),
-            last=?blocks.last().map(|b| b.recovered_block.num_hash()),
-            "storing range of blocks",
-        );
-
-        // insert blocks to db
-        let provider_rw = self.inner.blockchain_provider.database_provider_rw()?;
-        provider_rw.save_blocks(blocks.clone(), reth_provider::SaveBlocksMode::Full)?;
-        // advance the finalized/safe markers in the same transaction as the blocks: every
-        // canonical block comes from committed consensus output, so the last saved block is
-        // final by construction, and the single commit leaves no crash window where the
-        // persisted tip outruns the marker
-        if let Some(last) = blocks.last() {
-            let last_number = last.recovered_block.number;
-            provider_rw.save_finalized_block_number(last_number)?;
-            provider_rw.save_safe_block_number(last_number)?;
-        }
-        provider_rw.commit()?;
-
-        // process update
-        //
-        // see reth::EngineApiTreeHandler::on_canonical_chain_update
-        let chain_update = NewCanonicalChain::Commit { new: blocks };
-        let canonical_head = chain_update.tip();
-        let (epoch, round) =
-            deconstruct_nonce(<FixedBytes<8> as Into<u64>>::into(canonical_head.nonce));
-        info!(
-            target: "engine",
-            "canonical head for epoch {:?} round {:?}: {:?} - {:?}",
-            epoch,
-            round,
-            canonical_head.number,
-            canonical_head.hash(),
-        );
-
-        if let Some((leader_round, consensus_num_hash, engine_update_tx)) = engine_update {
-            engine_update_tx
-                .blocking_send((
-                    leader_round,
-                    consensus_num_hash,
-                    Some(canonical_head.clone_sealed_header()),
-                ))
-                .map_err(|e| {
-                    error!(target: "engine", ?e, "engine update channel send failed");
-                    TnRethError::EngineUpdateChannelClosed
-                })?;
-        }
-
-        // broadcast canonical update
-        let notification = chain_update.to_chain_notification();
-        self.canonical_in_memory_state().notify_canon_state(notification);
-
-        Ok(())
-    }
-
-    /// Look up and return the sealed header for hash.
-    pub fn sealed_header_by_hash(&self, hash: B256) -> TnRethResult<Option<SealedHeader>> {
-        Ok(self.inner.blockchain_provider.sealed_header_by_hash(hash)?)
-    }
-
-    /// Look up and return the sealed header for block number.
-    pub fn sealed_header_by_number(&self, number: u64) -> TnRethResult<Option<SealedHeader>> {
-        Ok(self.inner.blockchain_provider.database_provider_ro()?.sealed_header(number)?)
-    }
-
-    /// Look up and return the sealed block for number.
-    pub fn sealed_block_by_number(&self, number: u64) -> TnRethResult<Option<SealedBlock>> {
-        Ok(self
-            .inner
-            .blockchain_provider
-            .sealed_block_with_senders(
-                BlockHashOrNumber::Number(number),
-                TransactionVariant::NoHash,
-            )?
-            .map(|b| b.clone_sealed_block()))
-    }
-
-    /// Look up and return the sealed header (with senders) for hash.
-    pub fn sealed_block_with_senders(
-        &self,
-        id: BlockHashOrNumber,
-    ) -> TnRethResult<Option<BlockWithSenders>> {
-        Ok(self
-            .inner
-            .blockchain_provider
-            .sealed_block_with_senders(id, TransactionVariant::NoHash)?)
-    }
-
-    /// Return the blocks with senders for a range of block numbers.
-    pub fn block_with_senders_range(
-        &self,
-        range: RangeInclusive<BlockNumber>,
-    ) -> TnRethResult<Vec<BlockWithSenders>> {
-        Ok(self.inner.blockchain_provider.block_with_senders_range(range)?)
-    }
-
-    /// Return the blocks for a range of block numbers.
-    pub fn blocks_for_range(
-        &self,
-        range: RangeInclusive<BlockNumber>,
-    ) -> TnRethResult<Vec<SealedHeader>> {
-        Ok(self.inner.blockchain_provider.sealed_headers_range(range)?)
-    }
-
-    /// Build a [`Chain`] from a historical block in the database.
-    ///
-    /// Reads the block with recovered senders and its receipts, then constructs
-    /// a `Chain` suitable for ExEx replay notifications.
-    ///
-    /// Returns `None` if the block does not exist in the database.
-    ///
-    /// # Replay fidelity
-    ///
-    /// The returned `Chain` carries an **empty `BundleState`**: account/storage
-    /// diffs are already committed to the DB at execution time and are not
-    /// re-derived here. ExExes that need historical state diffs must read them
-    /// from the provider by block number. Live `ChainExecuted` notifications
-    /// (from `finish_executing_output`) *do* carry the full `BundleState`.
-    pub fn replay_block_as_chain(
-        &self,
-        block_number: BlockNumber,
-    ) -> TnRethResult<Option<Arc<reth_provider::Chain>>> {
-        // Read block with senders
-        let Some(block) = self.inner.blockchain_provider.sealed_block_with_senders(
-            BlockHashOrNumber::Number(block_number),
-            TransactionVariant::NoHash,
-        )?
-        else {
-            return Ok(None);
-        };
-
-        // Read receipts for this block. The block exists (read above), so missing
-        // receipts indicate a DB inconsistency; an empty block legitimately
-        // returns `Some(vec![])`. Treat `None` as an error rather than silently
-        // yielding an empty receipt set (which would make a non-empty block look
-        // empty to a stateful indexer).
-        let receipts = self
-            .inner
-            .blockchain_provider
-            .receipts_by_block(BlockHashOrNumber::Number(block_number))?
-            .ok_or(TnRethError::ReplayReceiptsMissing(block_number))?;
-
-        // Construct a minimal ExecutionOutcome with receipts only (no bundle state)
-        let execution_outcome = ExecutionOutcome::new(
-            Default::default(), // empty BundleState — state already committed to DB
-            vec![receipts],
-            block_number,
-            Vec::new(), // no requests
-        );
-
-        Ok(Some(Arc::new(Chain::new(vec![block], execution_outcome, Default::default()))))
-    }
-
-    /// Return the head header from the reth db.
-    pub fn lookup_head(&self) -> TnRethResult<SealedHeader> {
-        let head = self.inner.node_config.lookup_head(&self.inner.blockchain_provider)?;
-        let header = self
-            .inner
-            .blockchain_provider
-            .sealed_header(head.number)?
-            .expect("Failed to retrieve sealed header from head's block number");
-        Ok(header)
-    }
-
-    /// If a dubug max round is set then return it.
-    pub fn get_debug_max_round(&self) -> Option<u64> {
-        self.inner.node_config.debug.max_block
-    }
-
-    /// Helper to get the gas price based on the provider's latest header.
-    pub fn get_gas_price(&self) -> TnRethResult<u128> {
-        let header = self.lookup_head()?;
-        Ok(header.next_block_base_fee(BaseFeeParams::ethereum()).unwrap_or_default().into())
-    }
-
-    /// Return the execution header for hash if available.
-    pub fn header(&self, hash: B256) -> TnRethResult<Option<ExecHeader>> {
-        Ok(self.inner.blockchain_provider.header(hash)?)
-    }
-
-    /// Return the execution header for block number if available.
-    pub fn header_by_number(&self, block_num: u64) -> TnRethResult<Option<ExecHeader>> {
-        Ok(self.inner.blockchain_provider.database_provider_ro()?.header_by_number(block_num)?)
-    }
-
-    /// Return the finalized header, sealed with its hash, if available.
-    ///
-    /// Number and hash come from one logical read (the finalized num/hash pair, then the header
-    /// looked up by that hash), so callers can pin block ranges, epoch classification, and state
-    /// reads to this single header without consulting a second source (see `catchup_accumulator`
-    /// and the epoch-entry base-fee seeding in the epoch manager, which pair this with
-    /// [`Self::epoch_state_at_header`]).
-    pub fn finalized_header(&self) -> TnRethResult<Option<SealedHeader>> {
-        let Some(finalized_num_hash) = self.finalized_block_num_hash()? else {
-            return Ok(None);
-        };
-        self.sealed_header_by_hash(finalized_num_hash.hash)
-    }
-
-    /// Return the latest canonical block number.
-    pub fn last_block_number(&self) -> TnRethResult<u64> {
-        Ok(self.inner.blockchain_provider.database_provider_ro()?.last_block_number().unwrap_or(0))
-    }
-
-    /// Return the block number and hash for the current canonical tip.
-    ///
-    /// This checks the canonical-in-memory-state.
-    pub fn canonical_tip(&self) -> SealedHeader {
-        self.inner.blockchain_provider.canonical_in_memory_state().get_canonical_head()
-    }
-
-    /// If available return the finalized block number and hash.
-    ///
-    /// This checks the canonical-in-memory-state.
-    pub fn finalized_block_num_hash(&self) -> TnRethResult<Option<BlockNumHash>> {
-        Ok(self.inner.blockchain_provider.finalized_block_num_hash()?)
-    }
-
-    /// Returns the block number of the last finialized block.
-    pub fn last_finalized_block_number(&self) -> TnRethResult<u64> {
-        Ok(self
-            .inner
-            .blockchain_provider
-            .database_provider_ro()?
-            .last_finalized_block_number()?
-            .unwrap_or(0))
-    }
-
-    /// Return the block number and hash of the finalized block on node startup.
-    ///
-    /// This method adds additional fallbacks to ensure genesis is used when the network is starting
-    /// because the genesis block is not initialized as `finalized`. Nodes that start on genesis
-    /// will resync with the network if it exists.
-    pub fn finalized_block_hash_number_for_startup(&self) -> TnRethResult<BlockHashNumber> {
-        let hash = self
-            .inner
-            .blockchain_provider
-            .finalized_block_hash()?
-            .unwrap_or_else(|| self.inner.node_config.chain.sealed_genesis_header().hash());
-        let number = self.inner.blockchain_provider.finalized_block_number()?.unwrap_or_default();
-        Ok(BlockHashNumber { hash, number })
-    }
-
-    /// Build and return the RPC server for the instance.
-    /// This probably needs better abstraction.
-    pub fn get_rpc_server(
-        &self,
-        transaction_pool: WorkerTxPool,
-        network: WorkerNetwork,
-        other: impl Into<Methods>,
-    ) -> RpcServer {
-        let transaction_pool: EthTransactionPool<
-            BlockchainProvider<TelcoinNode>,
-            DiskFileBlobStore,
-            TnEvmConfig,
-        > = transaction_pool.into();
-        let tn_execution = Arc::new(TNExecution);
-        let rpc_builder = RpcModuleBuilder::default()
-            .with_provider(self.inner.blockchain_provider.clone())
-            .with_pool(transaction_pool.clone())
-            .with_network(network.clone())
-            .with_executor(Box::new(self.inner.task_spawner.clone()))
-            .with_evm_config(self.inner.evm_config.clone())
-            .with_consensus(tn_execution.clone());
-
-        let modules_config = self.inner.node_config.rpc.transport_rpc_module_config();
-        let eth_api = EthApi::builder(
-            self.inner.blockchain_provider.clone(),
-            transaction_pool,
-            network,
-            self.inner.evm_config.clone(),
-        )
-        .build();
-
-        let engine_events = reth_tokio_util::EventSender::default();
-        let mut server = rpc_builder.build(modules_config, eth_api, engine_events);
-        if let Err(e) = server.merge_configured(other) {
-            tracing::error!(target: "tn::execution", "Error merging TN rpc module: {e:?}");
-        }
-
-        server
-    }
-
-    /// Start running the RPC server for this instance.
-    pub async fn start_rpc(&self, server: &RpcServer) -> TnRethResult<RpcServerHandle> {
-        let server_config = self.inner.node_config.rpc.rpc_server_config();
-        Ok(server_config.start(server).await?)
-    }
-
-    /// Provide the state for the latest block in this instance.
-    pub fn latest(&self) -> TnRethResult<StateProviderBox> {
-        Ok(self.inner.blockchain_provider.latest()?)
-    }
-
-    /// Look up a transaction by hash, returning it with its sender and block metadata
-    /// (block number/hash, index in block, base fee, block timestamp).
-    pub fn transaction_by_hash_with_meta(
-        &self,
-        hash: TxHash,
-    ) -> TnRethResult<Option<(Recovered<TransactionSigned>, TransactionMeta)>> {
-        let Some((tx, meta)) =
-            self.inner.blockchain_provider.transaction_by_hash_with_meta(hash)?
-        else {
-            return Ok(None);
-        };
-
-        // prefer the sender persisted at execution time (two point reads, no ecrecover)
-        let sender = match self.inner.blockchain_provider.transaction_id(hash)? {
-            Some(id) => self.inner.blockchain_provider.transaction_sender(id)?,
-            None => None,
-        };
-        let sender = match sender {
-            Some(sender) => sender,
-            None => tx.recover_signer()?,
-        };
-
-        Ok(Some((Recovered::new_unchecked(tx, sender), meta)))
-    }
-
-    /// Return the receipt for a transaction hash, if the transaction has been mined.
-    ///
-    /// `Receipt::cumulative_gas_used` is cumulative within the block; use
-    /// [`Self::receipt_by_hash_with_gas_used`] for this transaction's own gas.
-    pub fn receipt_by_hash(&self, hash: TxHash) -> TnRethResult<Option<Receipt>> {
-        Ok(self.inner.blockchain_provider.receipt_by_hash(hash)?)
-    }
-
-    /// Return the receipt for a transaction hash together with the gas used by that
-    /// transaction alone (the delta of cumulative gas vs. the previous receipt in the block).
-    pub fn receipt_by_hash_with_gas_used(
-        &self,
-        hash: TxHash,
-    ) -> TnRethResult<Option<(Receipt, u64)>> {
-        let Some(id) = self.inner.blockchain_provider.transaction_id(hash)? else {
-            return Ok(None);
-        };
-        let Some(receipt) = self.inner.blockchain_provider.receipt(id)? else {
-            return Ok(None);
-        };
-        let block = self
-            .inner
-            .blockchain_provider
-            .block_by_transaction_id(id)?
-            .ok_or(ProviderError::TransactionNotFound(id.into()))?;
-        let indices = self
-            .inner
-            .blockchain_provider
-            .block_body_indices(block)?
-            .ok_or(ProviderError::BlockBodyIndicesNotFound(block))?;
-
-        let gas_used = if id == indices.first_tx_num() {
-            receipt.cumulative_gas_used
-        } else {
-            let prev = self
-                .inner
-                .blockchain_provider
-                .receipt(id - 1)?
-                .ok_or(ProviderError::ReceiptNotFound((id - 1).into()))?;
-            receipt.cumulative_gas_used.saturating_sub(prev.cumulative_gas_used)
-        };
-
-        Ok(Some((receipt, gas_used)))
-    }
-
-    /// Return all receipts for a block by hash or number.
-    ///
-    /// `Ok(None)` means the block is unknown; an empty vec means the block has no transactions.
-    pub fn receipts_by_block(
-        &self,
-        block: BlockHashOrNumber,
-    ) -> TnRethResult<Option<Vec<Receipt>>> {
-        Ok(self.inner.blockchain_provider.receipts_by_block(block)?)
-    }
-
-    /// Total number of transactions ever mined on the canonical chain.
-    ///
-    /// The latest transaction's [`TxNumber`] is `total - 1`; serve "latest N, newest-first"
-    /// pages by reading descending [`TxNumber`] ranges via
-    /// [`Self::transactions_by_tx_range_with_meta`].
-    pub fn total_transactions(&self) -> TnRethResult<u64> {
-        let tip = self.last_block_number()?;
-        Ok(self
-            .inner
-            .blockchain_provider
-            .block_body_indices(tip)?
-            .map(|indices| indices.next_tx_num())
-            .unwrap_or(0))
-    }
-
-    /// Return the transactions in a chain-wide [`TxNumber`] range (inclusive), each with its
-    /// sender and containing-block metadata.
-    ///
-    /// Cost scales with the number of transactions returned, not the number of blocks the
-    /// range spans: empty blocks are never visited. Ranges extending past the newest
-    /// transaction are clamped to what exists.
-    pub fn transactions_by_tx_range_with_meta(
-        &self,
-        range: RangeInclusive<TxNumber>,
-    ) -> TnRethResult<Vec<TxFeedEntry>> {
-        if range.is_empty() {
-            return Ok(Vec::new());
-        }
-        let start = *range.start();
-        let txs = self.inner.blockchain_provider.transactions_by_tx_range(range.clone())?;
-        if txs.is_empty() {
-            return Ok(Vec::new());
-        }
-        let mut senders = self.inner.blockchain_provider.senders_by_tx_range(range)?;
-        if senders.len() != txs.len() {
-            // defensive: persistence always writes senders alongside transactions, so a
-            // misaligned read should be impossible — recover in parallel if it happens
-            senders = txs
-                .par_iter()
-                .map(|tx| tx.recover_signer().map_err(TnRethError::from))
-                .collect::<TnRethResult<Vec<_>>>()?;
-        }
-
-        let mut entries = Vec::with_capacity(txs.len());
-        // cached (block_number, first_tx_num, next_tx_num, block_hash, timestamp) for the block
-        // containing the previous transaction; refreshed only when the cursor crosses into the
-        // next non-empty block, so each non-empty block costs three point reads and empty
-        // blocks are never visited (the `TransactionBlocks` seek skips them by construction).
-        let mut block_ctx: Option<(BlockNumber, TxNumber, TxNumber, B256, u64)> = None;
-        for (offset, (tx, sender)) in txs.into_iter().zip(senders).enumerate() {
-            let tx_number = start + offset as u64;
-            let (block_number, first_tx_num, _, block_hash, timestamp) = match block_ctx {
-                Some(ctx) if tx_number < ctx.2 => ctx,
-                _ => {
-                    let block = self
-                        .inner
-                        .blockchain_provider
-                        .block_by_transaction_id(tx_number)?
-                        .ok_or(ProviderError::TransactionNotFound(tx_number.into()))?;
-                    let indices = self
-                        .inner
-                        .blockchain_provider
-                        .block_body_indices(block)?
-                        .ok_or(ProviderError::BlockBodyIndicesNotFound(block))?;
-                    let header = self
-                        .inner
-                        .blockchain_provider
-                        .sealed_header(block)?
-                        .ok_or(ProviderError::HeaderNotFound(block.into()))?;
-                    let ctx = (
-                        block,
-                        indices.first_tx_num(),
-                        indices.next_tx_num(),
-                        header.hash(),
-                        header.timestamp,
-                    );
-                    block_ctx = Some(ctx);
-                    ctx
-                }
-            };
-            entries.push(TxFeedEntry {
-                tx_number,
-                transaction: Recovered::new_unchecked(tx, sender),
-                block_number,
-                block_hash,
-                timestamp,
-                index: tx_number.saturating_sub(first_tx_num),
-            });
-        }
-
-        Ok(entries)
-    }
-
-    /// Return balance, nonce, and code hash for an account at the latest canonical state.
-    pub fn retrieve_account(&self, address: &Address) -> TnRethResult<Option<Account>> {
-        Ok(self.inner.blockchain_provider.basic_account(address)?)
-    }
-
-    /// Return the contract bytecode deployed at `address` at the latest canonical state
-    /// (`eth_getCode` equivalent). `None` for EOAs and unknown accounts.
-    pub fn account_code(&self, address: &Address) -> TnRethResult<Option<Bytes>> {
-        Ok(self.latest()?.account_code(address)?.map(|code| code.original_bytes()))
-    }
-
-    /// Create a new temp RethEnv using a specified chain spec.
-    pub fn new_for_temp_chain<P: AsRef<Path>>(
-        chain: Arc<RethChainSpec>,
-        db_path: P,
-        task_manager: &TaskManager,
-        rewards: Option<RewardsCounter>,
-    ) -> eyre::Result<Self> {
-        /// MDBX map-size ceiling for throwaway temp-chain envs. reth defaults to 8 TB per
-        /// environment; `cargo test` runs a test binary as threads in ONE process, so N
-        /// concurrent 8 TB virtual reservations exhaust the address space and MDBX aborts
-        /// env-open with ENOMEM ("Cannot allocate memory (12)"). A temp chain never holds a
-        /// full node's history, so a small ceiling is safe and lets many envs coexist.
-        const TEMP_CHAIN_DB_MAX_SIZE: usize = 2 * 1024 * 1024 * 1024; // 2 GiB
-        /// Grow the temp DB file in small increments so throwaway DBs stay tiny on disk
-        /// (reth pairs its 8 TB default with a 4 GiB step; mirror reth's `test()` 4 MiB step).
-        const TEMP_CHAIN_DB_GROWTH_STEP: usize = 4 * 1024 * 1024; // 4 MiB
-
-        // `NodeConfig::default` reads reth's process-wide pool defaults through
-        // `TxPoolArgs::default`, which locks them. Seed first so a temp chain built before any
-        // command is parsed does not fix the per-sender slot default at reth's 16 for the rest of
-        // the process, which would then also apply to every node parsed later.
-        init_txpool_defaults();
-
-        let node_config = NodeConfig {
-            datadir: DatadirArgs {
-                datadir: MaybePlatformPath::from(db_path.as_ref().to_path_buf()),
-                // default static path should resolve to: `DEFAULT_ROOT_DIR/<CHAIN_ID>/static_files`
-                static_files_path: None,
-                rocksdb_path: None,
-                pprof_dumps_path: None,
-            },
-            chain,
-            // Bound the MDBX geometry for throwaway temp chains (see const docs).
-            db: DatabaseArgs {
-                max_size: Some(TEMP_CHAIN_DB_MAX_SIZE),
-                growth_step: Some(TEMP_CHAIN_DB_GROWTH_STEP),
-                ..Default::default()
-            },
-            ..NodeConfig::default()
-        };
-        let reth_config = RethConfig(node_config);
-        let database = Self::new_database(&reth_config, db_path)?;
-        Self::new(&reth_config, task_manager, database, None, rewards.unwrap_or_default())
-    }
-
-    /// Convenience method for compiling storage and bytecode to include consensus registry
-    /// configuration in genesis.
-    pub fn create_consensus_registry_genesis_accounts(
-        validators: Vec<NodeInfo>,
-        genesis: Genesis,
-        initial_stake_config: ConsensusRegistry::StakeConfig,
-        owner_address: Address,
-        worker_configs: Vec<(u8, u64)>,
-    ) -> eyre::Result<Genesis> {
-        // create temporary reth env for execution
-        let tmp_chain: Arc<RethChainSpec> = Arc::new(genesis.clone().into());
-        let task_manager = TaskManager::new("Temp Task Manager");
-        let tmp_dir = TempDir::new().unwrap();
-        let reth_env =
-            RethEnv::new_for_temp_chain(tmp_chain.clone(), tmp_dir.path(), &task_manager, None)?;
-
-        let state = StateProviderDatabase::new(reth_env.latest()?);
-        let mut cached_reads = CachedReads::default();
-        let mut db = State::builder()
-            .with_database(cached_reads.as_db_mut(state))
-            .with_bundle_update()
-            .build();
-
-        // The BLS proof-of-possession precompile is registered by the EVM factory at
-        // `BLS_G1_PRECOMPILE_ADDRESS`, so there is nothing to deploy at genesis. The registry
-        // constructor's PoP verification staticcalls this address and is served natively by the
-        // precompile, which shares the exact `blst` verification the consensus layer uses to
-        // produce the proofs (so the on-chain check cannot drift from the off-chain
-        // encoding).
-        let blsg1_address = BLS_G1_PRECOMPILE_ADDRESS;
-
-        // prepare registry deployment
-        let (validators, (bls_pubkeys, proofs)): (Vec<_>, (Vec<_>, Vec<_>)) = validators
-            .iter()
-            .map(|v| {
-                let validator = ConsensusRegistry::ValidatorInfo {
-                    validatorAddress: v.execution_address,
-                    activationEpoch: 0,
-                    exitEpoch: 0,
-                    currentStatus: ConsensusRegistry::ValidatorStatus::Active,
-                    isRetired: false,
-                    stakeVersion: 0,
-                    region: 0,
-                };
-                let bls_pubkey: tn_types::Bytes = v.bls_public_key.to_bytes().into();
-                let proof = ConsensusRegistry::ProofOfPossession {
-                    signature: v.proof_of_possession.to_bytes().into(),
-                };
-
-                (validator, (bls_pubkey, proof))
-            })
-            .unzip();
-
-        let total_stake_balance = initial_stake_config
-            .stakeAmount
-            .checked_mul(U256::from(validators.len()))
-            .ok_or_eyre("Failed to calculate total stake for consensus registry at genesis")?;
-        debug!(target: "engine", ?initial_stake_config, "calling constructor for consensus registry");
-
-        let constructor_args = ConsensusRegistry::constructorCall {
-            genesisConfig_: initial_stake_config,
-            initialValidators_: validators,
-            blsPubkeys_: bls_pubkeys,
-            proofsOfPossession: proofs,
-            owner_: owner_address,
-        }
-        .abi_encode();
-
-        // generate calldata for creation
-        let registry_initcode_binding =
-            Self::fetch_value_from_json_str(CONSENSUS_REGISTRY_JSON, Some("bytecode.object"))?;
-        let registry_initcode_str =
-            registry_initcode_binding.as_str().ok_or_eyre("invalid registry json")?;
-        // The registry calls the BLS precompile directly at `BLS_G1_ADDRESS` (no linked library),
-        // so its bytecode carries no link placeholder; deploy it as-is. Guard against a
-        // stale artifact that still contains an unresolved `__$..$__` placeholder.
-        if registry_initcode_str.contains("__$") {
-            eyre::bail!(
-                "ConsensusRegistry initcode has an unresolved library link placeholder; regenerate tn-contracts artifacts"
-            );
-        }
-        let mut create_registry = hex::decode(registry_initcode_str)?;
-        create_registry.extend(constructor_args);
-
-        // after adding bls proof of possession, registry precompile exceeds size limit so disable
-        // it for tmp chain
-        let mut tmp_evm_no_eip170 =
-            reth_env.inner.evm_config.evm_env(&tmp_chain.sealed_genesis_header())?;
-        tmp_evm_no_eip170.cfg_env.limit_contract_code_size = Some(0x12000000);
-
-        // deploy registry now that it can use the previously deployed blsg1 lib
-        let tmp_registry_address = {
-            let mut tn_evm =
-                reth_env.inner.evm_config.evm_factory().create_evm(&mut db, tmp_evm_no_eip170);
-            let ResultAndState { result, state } =
-                tn_evm.transact_pre_genesis_create(owner_address, create_registry.into())?;
-            debug!(target: "engine", "create consensus registry result:\n{:#?}", result);
-            Self::ensure_pre_genesis_create_success("ConsensusRegistry", &result)?;
-
-            tn_evm.db_mut().commit(state);
-
-            // With BlsG1 now a precompile (no genesis deploy), the registry is the owner's first
-            // create tx on the tmp chain.
-            owner_address.create(0)
-        };
-
-        // deploy WorkerConfigs contract
-        let tmp_worker_configs_address = {
-            let mut tn_evm = reth_env.inner.evm_config.evm_factory().create_evm(
-                &mut db,
-                reth_env.inner.evm_config.evm_env(&tmp_chain.sealed_genesis_header())?,
-            );
-
-            let (strategies, values): (Vec<u8>, Vec<u64>) = worker_configs.iter().copied().unzip();
-            let datas = vec![0u128; strategies.len()];
-            let constructor_args =
-                WorkerConfigs::constructorCall { strategies, values, datas, owner_: owner_address }
-                    .abi_encode();
-
-            let initcode_binding =
-                Self::fetch_value_from_json_str(WORKER_CONFIGS_JSON, Some("bytecode.object"))?;
-            let initcode =
-                hex::decode(initcode_binding.as_str().ok_or_eyre("invalid worker configs json")?)?;
-            let mut create_worker_configs = initcode;
-            create_worker_configs.extend(constructor_args);
-
-            let ResultAndState { result, state } =
-                tn_evm.transact_pre_genesis_create(owner_address, create_worker_configs.into())?;
-            debug!(target: "engine", "create worker configs result:\n{:#?}", result);
-            Self::ensure_pre_genesis_create_success("WorkerConfigs", &result)?;
-
-            tn_evm.db_mut().commit(state);
-
-            // Second create tx on tmp chain (registry at nonce 0, worker configs at nonce 1).
-            owner_address.create(1)
-        };
-
-        // execute the transactions to get final bundle state
-        db.merge_transitions(BundleRetention::PlainState);
-        let BundleState { state, contracts, reverts, state_size, reverts_size } = db.take_bundle();
-
-        debug!(target: "engine", "contracts:\n{:#?}", contracts);
-        debug!(target: "engine", "reverts:\n{:#?}", reverts);
-        debug!(target: "engine", "state_size:{:#?}", state_size);
-        debug!(target: "engine", "reverts_size:{:#?}", reverts_size);
-
-        // construct real genesis using known values & tmp chain storage result
-        let tmp_registry_storage = state.get(&tmp_registry_address).map(|account| {
-            account.storage.iter().map(|(k, v)| ((*k).into(), v.present_value.into())).collect()
-        });
-        let registry_runtimecode_binding = Self::fetch_value_from_json_str(
-            CONSENSUS_REGISTRY_JSON,
-            Some("deployedBytecode.object"),
-        )?;
-        let registry_runtimecode_str =
-            registry_runtimecode_binding.as_str().ok_or_eyre("invalid registry json")?;
-        let registry_runtimecode = hex::decode(registry_runtimecode_str)?;
-
-        let tmp_worker_configs_storage = state.get(&tmp_worker_configs_address).map(|account| {
-            account.storage.iter().map(|(k, v)| ((*k).into(), v.present_value.into())).collect()
-        });
-        let worker_configs_runtimecode_binding =
-            Self::fetch_value_from_json_str(WORKER_CONFIGS_JSON, Some("deployedBytecode.object"))?;
-        let worker_configs_runtimecode = hex::decode(
-            worker_configs_runtimecode_binding
-                .as_str()
-                .ok_or_eyre("invalid worker configs json")?,
-        )?;
-
-        let issuance_json_binding =
-            Self::fetch_value_from_json_str(ISSUANCE_JSON, Some("deployedBytecode.object"))?;
-        let issuance_runtimecode =
-            hex::decode(issuance_json_binding.as_str().ok_or_eyre("invalid issuance json")?)?;
-        let genesis = genesis.extend_accounts([
-            // The BLS proof-of-possession precompile lives at `blsg1_address`
-            // (`BLS_G1_PRECOMPILE_ADDRESS`). Mirror the TEL precompile and give it a single `0xfe`
-            // (INVALID) byte of code so the account is non-empty (never state-pruned) and any call
-            // that bypasses precompile dispatch reverts instead of succeeding against an EOA.
-            (
-                blsg1_address,
-                GenesisAccount::default().with_code(Some(PRECOMPILE_GENESIS_BYTECODE.into())),
-            ),
-            (
-                CONSENSUS_REGISTRY_ADDRESS,
-                GenesisAccount::default()
-                    .with_balance(U256::from(total_stake_balance))
-                    .with_code(Some(registry_runtimecode.into()))
-                    .with_storage(tmp_registry_storage),
-            ),
-            (
-                ISSUANCE_ADDRESS,
-                GenesisAccount::default().with_code(Some(issuance_runtimecode.into())),
-            ),
-            (
-                WORKER_CONFIGS_ADDRESS,
-                GenesisAccount::default()
-                    .with_code(Some(worker_configs_runtimecode.into()))
-                    .with_storage(tmp_worker_configs_storage),
-            ),
-        ]);
-
-        Ok(genesis)
-    }
-
-    /// Bail unless a pre-genesis constructor transaction succeeded.
-    ///
-    /// Pre-genesis creates are committed straight into genesis storage. An unchecked
-    /// Revert/Halt would still ship the contract's runtime code but with EMPTY storage
-    /// (e.g. an ownerless, zero-worker `WorkerConfigs`), which downstream fail-open reads
-    /// mask as defaults — so the ceremony must fail loudly here instead.
-    fn ensure_pre_genesis_create_success(
-        contract: &str,
-        result: &ExecutionResult,
-    ) -> eyre::Result<()> {
-        match result {
-            ExecutionResult::Success { .. } => Ok(()),
-            ExecutionResult::Revert { output, .. } => {
-                let reason = alloy::sol_types::decode_revert_reason(output)
-                    .unwrap_or_else(|| "<undecodable revert reason>".to_string());
-                eyre::bail!(
-                    "{contract} constructor reverted during pre-genesis create: {reason} (revert output: {output})"
-                )
-            }
-            ExecutionResult::Halt { reason, gas_used } => eyre::bail!(
-                "{contract} constructor halted during pre-genesis create: {reason:?} (gas used: {gas_used})"
-            ),
-        }
-    }
-
-    /// Fetches json info from the given string
-    ///
-    /// If a key is specified, return the corresponding nested object.
-    /// Otherwise return the entire JSON
-    /// With a generic this could be adjused to handle YAML also
-    pub fn fetch_value_from_json_str(json_content: &str, key: Option<&str>) -> eyre::Result<Value> {
-        let json: Value = serde_json::from_str(json_content)?;
-        let result = match key {
-            Some(path) => {
-                let key: Vec<&str> = path.split('.').collect();
-                let mut current_value = &json;
-                for &k in &key {
-                    current_value =
-                        current_value.get(k).ok_or_else(|| eyre::eyre!("key '{}' not found", k))?;
-                }
-                current_value.clone()
-            }
-            None => json,
-        };
-
-        Ok(result)
-    }
-
-    /// Read the latest committee and epoch information from the [ConsensusRegistry] on-chain.
-    ///
-    /// The protocol needs the BLS pubkey for the authorities.
-    /// - get current epoch info
-    /// - getValidator token id by address
-    /// - getValidator info by token id
-    ///
-    /// The committee arrays this returns mutate mid-epoch: a governance `burn` swap-and-pops the
-    /// ejected validator out of the CURRENT epoch's stored committees immediately, so tip reads
-    /// before and after the burn disagree. Epoch-scoped consensus reads (committee construction,
-    /// rewards seeding, quorum) must use [`Self::epoch_state_at_epoch_start`] instead; the tip
-    /// read remains correct for point-in-time queries.
-    pub fn epoch_state_from_canonical_tip(&self) -> eyre::Result<EpochState> {
-        let canonical_tip = self.canonical_tip();
-        debug!(target: "engine", ?canonical_tip, "retrieving epoch state from canonical tip");
-        self.epoch_state_at_header(&canonical_tip)
-    }
-
-    /// Read the committee and epoch information from the [ConsensusRegistry] at `header`.
-    ///
-    /// The registry state, the EVM environment, and therefore the returned epoch, epoch info,
-    /// and committee all derive from this ONE header. Recovery paths that scan
-    /// `epoch_info.blockHeight..=header.number` (catchup and epoch-entry base-fee seeding) rely
-    /// on this pin: reading the range start from a different header (e.g. the canonical tip)
-    /// could yield a silently empty range if finality ever lags the canonical tip.
-    pub fn epoch_state_at_header(&self, header: &SealedHeader) -> eyre::Result<EpochState> {
-        // create EVM with the state at the pinned header
-        let state_provider = self.inner.blockchain_provider.state_by_block_hash(header.hash())?;
-        let state = StateProviderDatabase::new(&state_provider);
-        let mut db = State::builder().with_database(state).with_bundle_update().build();
-        debug!(target: "engine", state=?db.bundle_state, hashes=?db.block_hashes, "retrieving epoch state at header");
-        let mut tn_evm = self
-            .inner
-            .evm_config
-            .evm_factory()
-            .create_evm(&mut db, self.inner.evm_config.evm_env(header)?);
-
-        // current epoch number
-        let epoch = self.get_current_epoch_number(&mut tn_evm)?;
-
-        // current epoch info
-        let epoch_info = self.get_current_epoch_info(&mut tn_evm)?;
-        debug!(target: "engine", ?epoch, ?epoch_info, "retrieved epoch info at header");
-
-        // retrieve closing timestamp for previous epoch
-        let epoch_start = self
-            .header_by_number(epoch_info.blockHeight.saturating_sub(1))?
-            .ok_or_eyre("failed to retrieve closing epoch information")?
-            .timestamp;
-
-        // retrieve the committee
-        let validators = self.get_committee_validators_by_epoch(epoch, &mut tn_evm)?;
-        let bls_pubkeys = self.get_committee_bls_pubkeys_by_epoch(epoch, &mut tn_evm)?;
-        let epoch_state = EpochState { epoch, epoch_info, validators, bls_pubkeys, epoch_start };
-        debug!(target: "engine", ?epoch_state, "returning epoch state at header");
-
-        Ok(epoch_state)
-    }
-
-    /// Read the CURRENT epoch's [`EpochState`] pinned to the epoch's start state — the previous
-    /// epoch's closing block (genesis for epoch 0) — returning the pin header alongside it.
-    ///
-    /// Bootstrapping reads the current epoch number and its
-    /// [`EpochInfo`](ConsensusRegistry::EpochInfo) at the canonical tip. This is deterministic at
-    /// ANY tip because `concludeEpoch` writes the epoch number and the `EpochInfo` scalars this
-    /// method consumes (`blockHeight`) exactly once at the boundary and never mid-epoch — only
-    /// the committee ARRAYS mutate mid-epoch (governance `burn` swap-and-pops immediately,
-    /// including the committee embedded in `EpochInfo`), which is why the full state read below
-    /// is pinned instead of taken at the tip.
-    ///
-    /// The pin: `epoch_info.blockHeight` is the entered epoch's first block, so `blockHeight - 1`
-    /// is the previous epoch's closing block — the block whose execution ran `concludeEpoch` and
-    /// seated the entered committee. For epoch 0 the pin is genesis, whose execution seeds the
-    /// registry: genesis state IS epoch 0's start state.
-    ///
-    /// Every field of the returned state derives from the previous epoch's closing block, so any
-    /// node entering (or re-entering) the same epoch — fresh boundary crossing, crash-restart
-    /// replay, or ModeChange re-entry, before or after a mid-epoch burn — derives an IDENTICAL
-    /// epoch view (committee membership included).
-    pub fn epoch_state_at_epoch_start(&self) -> eyre::Result<(EpochState, SealedHeader)> {
-        // boundary-written-once identity read: deterministic at any tip
-        let (epoch, epoch_info) = self.get_current_epoch_info_at_header(&self.canonical_tip())?;
-
-        let pin_header = if epoch == 0 {
-            self.sealed_header_by_number(0)?.ok_or_else(|| eyre::eyre!("missing genesis header"))?
-        } else {
-            let closing_number = epoch_info
-                .blockHeight
-                .checked_sub(1)
-                .ok_or_else(|| eyre::eyre!("current epoch {epoch} reports blockHeight 0"))?;
-            self.sealed_header_by_number(closing_number)?.ok_or_else(|| {
-                eyre::eyre!("missing closing block {closing_number} for current epoch {epoch}")
-            })?
-        };
-        debug!(
-            target: "engine",
-            ?epoch,
-            pin_number = pin_header.number,
-            pin_hash = ?pin_header.hash(),
-            "retrieving epoch state at epoch start"
-        );
-
-        let state = self.epoch_state_at_header(&pin_header)?;
-
-        // Tripwire: `concludeEpoch` executes INSIDE the closing block, so the registry at the
-        // closing header already reports the entered epoch. This can only fire if the registry's
-        // `blockHeight == closing block + 1` convention ever breaks — running a stale committee
-        // is a consensus-safety failure, so fail hard instead of returning the mismatched state.
-        if state.epoch != epoch {
-            error!(
-                target: "engine",
-                pin_number = pin_header.number,
-                pinned_epoch = state.epoch,
-                tip_epoch = epoch,
-                "epoch-start pin and canonical tip disagree on the current epoch"
-            );
-            return Err(eyre::eyre!(
-                "epoch state pinned to block {} reports epoch {} but the canonical tip reports \
-                 epoch {epoch}",
-                pin_header.number,
-                state.epoch
-            ));
-        }
-
-        Ok((state, pin_header))
-    }
-
-    /// Read the latest committee and epoch information from the [ConsensusRegistry] on-chain.
-    pub fn validators_for_epoch(
-        &self,
-        epoch: u32,
-    ) -> eyre::Result<Vec<ConsensusRegistry::ValidatorInfo>> {
-        debug!(target: "engine", "retrieving validators for epoch {epoch}");
-        let calldata = ConsensusRegistry::getCommitteeValidatorsCall { epoch }.abi_encode().into();
-        self.read_consensus_registry(calldata).map_err(Into::into)
-    }
-
-    /// Read the committee validators for the provided epoch from the [ConsensusRegistry], pinned
-    /// to the state of the block identified by `block_hash`.
-    ///
-    /// Unlike [`Self::validators_for_epoch`], which reads the mutable canonical tip, every node
-    /// issuing this read at the same block decodes the identical committee — even after a
-    /// mid-epoch governance `burn` swap-and-pops the stored committee arrays.
-    pub fn validators_for_epoch_at_block(
-        &self,
-        epoch: u32,
-        block_hash: B256,
-    ) -> eyre::Result<Vec<ConsensusRegistry::ValidatorInfo>> {
-        debug!(target: "engine", ?block_hash, "retrieving validators for epoch {epoch} at pinned block");
-        let header = self
-            .sealed_header_by_hash(block_hash)?
-            .ok_or_else(|| eyre::eyre!("sealed header not found for block hash {block_hash:?}"))?;
-        let calldata = ConsensusRegistry::getCommitteeValidatorsCall { epoch }.abi_encode().into();
-        self.read_consensus_registry_at_header(&header, calldata).map_err(Into::into)
-    }
-
-    /// Read the BLS pubkeys for the committee of the provided epoch from the [ConsensusRegistry]
-    /// on-chain.
-    pub fn bls_pubkeys_for_epoch(&self, epoch: u32) -> eyre::Result<Vec<alloy::primitives::Bytes>> {
-        let calldata = ConsensusRegistry::getCommitteeBlsPubkeysCall { epoch }.abi_encode().into();
-        self.read_consensus_registry(calldata).map_err(Into::into)
-    }
-
-    /// Read the BLS pubkeys for the committee of the provided epoch from the
-    /// [ConsensusRegistry], pinned to the state of the block identified by `block_hash`.
-    ///
-    /// Unlike [`Self::bls_pubkeys_for_epoch`], which reads the mutable canonical tip, every node
-    /// issuing this read at the same block decodes the identical key set — even after a
-    /// mid-epoch governance `burn` swap-and-pops the stored committee arrays.
-    pub fn bls_pubkeys_for_epoch_at_block(
-        &self,
-        epoch: u32,
-        block_hash: B256,
-    ) -> eyre::Result<Vec<alloy::primitives::Bytes>> {
-        let header = self
-            .sealed_header_by_hash(block_hash)?
-            .ok_or_else(|| eyre::eyre!("sealed header not found for block hash {block_hash:?}"))?;
-        let calldata = ConsensusRegistry::getCommitteeBlsPubkeysCall { epoch }.abi_encode().into();
-        self.read_consensus_registry_at_header(&header, calldata).map_err(Into::into)
-    }
-
-    /// Build an EVM at the canonical tip, execute a read-only [ConsensusRegistry] call, and
-    /// decode the returned data to `T`.
-    ///
-    /// Convenience wrapper over [`Self::read_consensus_registry_batch`] for the common
-    /// single-read case (one pinned EVM, one call).
-    pub fn read_consensus_registry<T>(&self, calldata: Bytes) -> EvmReadResult<T>
-    where
-        T: alloy::sol_types::SolValue,
-        T: From<
-            <<T as alloy::sol_types::SolValue>::SolType as alloy::sol_types::SolType>::RustType,
-        >,
-    {
-        self.read_consensus_registry_batch(vec![calldata])?.pop().ok_or_else(|| {
-            EvmReadError::Internal("consensus registry batch read returned no result".into())
-        })
-    }
-
-    /// Build an EVM at `header`'s state, execute a read-only [ConsensusRegistry] call, and
-    /// decode the returned data to `T`.
-    ///
-    /// Convenience wrapper over [`Self::read_consensus_registry_batch_at_header`] for the common
-    /// single-read case (one pinned EVM, one call).
-    pub fn read_consensus_registry_at_header<T>(
-        &self,
-        header: &SealedHeader,
-        calldata: Bytes,
-    ) -> EvmReadResult<T>
-    where
-        T: alloy::sol_types::SolValue,
-        T: From<
-            <<T as alloy::sol_types::SolValue>::SolType as alloy::sol_types::SolType>::RustType,
-        >,
-    {
-        self.read_consensus_registry_batch_at_header(header, vec![calldata])?.pop().ok_or_else(
-            || EvmReadError::Internal("consensus registry batch read returned no result".into()),
-        )
-    }
-
-    /// Build a single EVM at the canonical tip and execute several read-only [ConsensusRegistry]
-    /// calls against it, decoding each result to `T`.
-    ///
-    /// Thin specialization of [`Self::read_consensus_registry_batch_at_header`] that pins the
-    /// batch to the canonical tip.
-    pub fn read_consensus_registry_batch<T>(&self, calldatas: Vec<Bytes>) -> EvmReadResult<Vec<T>>
-    where
-        T: alloy::sol_types::SolValue,
-        T: From<
-            <<T as alloy::sol_types::SolValue>::SolType as alloy::sol_types::SolType>::RustType,
-        >,
-    {
-        let canonical_tip = self.canonical_tip();
-        debug!(target: "engine", ?canonical_tip, "reading consensus registry batch at canonical tip");
-        self.read_consensus_registry_batch_at_header(&canonical_tip, calldatas)
-    }
-
-    /// Build a single EVM at `header`'s state and execute several read-only [ConsensusRegistry]
-    /// calls against it, decoding each result to `T`.
-    ///
-    /// Every calldata in a batch must decode to the same Solidity type `T` (current caller: five
-    /// `getValidatorsInfo(status)` reads → `Vec<ValidatorInfo>`).
-    ///
-    /// All calls observe ONE pinned state snapshot, so a multi-call query (e.g. unioning
-    /// per-status validator sets) cannot straddle a block commit and double-count or drop a
-    /// validator that changes status between reads. Each call still runs under its own fresh 30M
-    /// gas budget (`transact_system_call`), so splitting a large query across calls keeps gas
-    /// bounded per call.
-    pub fn read_consensus_registry_batch_at_header<T>(
-        &self,
-        header: &SealedHeader,
-        calldatas: Vec<Bytes>,
-    ) -> EvmReadResult<Vec<T>>
-    where
-        T: alloy::sol_types::SolValue,
-        T: From<
-            <<T as alloy::sol_types::SolValue>::SolType as alloy::sol_types::SolType>::RustType,
-        >,
-    {
-        // Create EVM with the state at the pinned header.
-        //
-        // ARCHIVE-MODE ASSUMPTION: this node never constructs a pruner (`PruningArgs` are built
-        // with every field disabled and no `PrunerBuilder` exists in the repo), so
-        // `state_by_block_hash` always resolves fully indexed history. If pruning is ever
-        // enabled, reth's `HistoricalStateProvider` can hit a missing history shard and return
-        // `HistoryInfo::MaybeInPlainState`, silently falling back to TIP state for this
-        // "pinned" read — exactly the nondeterminism pinning exists to prevent. Revisit every
-        // pinned registry read before enabling pruning.
-        let state_provider = self
-            .inner
-            .blockchain_provider
-            .state_by_block_hash(header.hash())
-            .map_err(|e| EvmReadError::Internal(e.to_string()))?;
-        let state = StateProviderDatabase::new(&state_provider);
-        let mut db = State::builder().with_database(state).with_bundle_update().build();
-        let evm_env = self
-            .inner
-            .evm_config
-            .evm_env(header)
-            .map_err(|e| EvmReadError::Internal(e.to_string()))?;
-        let mut tn_evm = self.inner.evm_config.evm_factory().create_evm(&mut db, evm_env);
-
-        // reuse the one pinned EVM for every read; `call_consensus_registry` is non-committing,
-        // so each read sees the same base state.
-        calldatas
-            .into_iter()
-            .map(|calldata| self.call_consensus_registry(&mut tn_evm, calldata))
-            .collect()
-    }
-
-    /// Execute a read-only contract call at the canonical tip and return the raw ABI-encoded
-    /// output bytes.
-    ///
-    /// `eth_call`-like semantics: caller is the zero address, value 0, gas price 0, nonce and
-    /// base fee checks disabled, 30M gas; no state is committed. Callers decode the returned
-    /// bytes themselves (e.g. with `SolCall::abi_decode_returns`).
-    pub fn read_contract(&self, contract: Address, calldata: Bytes) -> EvmReadResult<Bytes> {
-        self.read_contract_inner(&self.canonical_tip(), Address::ZERO, contract, calldata)
-    }
-
-    /// Same as [`Self::read_contract`], but pinned to the state of the block identified by
-    /// `block_hash`.
-    pub fn read_contract_at_block(
-        &self,
-        block_hash: B256,
-        contract: Address,
-        calldata: Bytes,
-    ) -> EvmReadResult<Bytes> {
-        let header = self
-            .sealed_header_by_hash(block_hash)
-            .map_err(|e| EvmReadError::Internal(e.to_string()))?
-            .ok_or_else(|| {
-                EvmReadError::Internal(format!(
-                    "sealed header not found for block hash {block_hash:?}"
-                ))
-            })?;
-        self.read_contract_inner(&header, Address::ZERO, contract, calldata)
-    }
-
-    /// Build an EVM against `header`'s state and execute one read-only call, returning raw
-    /// output bytes on success and mapping Revert/Halt like the registry read path.
-    fn read_contract_inner(
-        &self,
-        header: &SealedHeader,
-        caller: Address,
-        contract: Address,
-        calldata: Bytes,
-    ) -> EvmReadResult<Bytes> {
-        let state_provider = self
-            .inner
-            .blockchain_provider
-            .state_by_block_hash(header.hash())
-            .map_err(|e| EvmReadError::Internal(e.to_string()))?;
-        let state = StateProviderDatabase::new(&state_provider);
-        let mut db = State::builder().with_database(state).with_bundle_update().build();
-        let evm_env = self
-            .inner
-            .evm_config
-            .evm_env(header)
-            .map_err(|e| EvmReadError::Internal(e.to_string()))?;
-        let mut tn_evm = self.inner.evm_config.evm_factory().create_evm(&mut db, evm_env);
-
-        let result = self
-            .read_state_on_chain(&mut tn_evm, caller, contract, calldata)
-            .map_err(|e| EvmReadError::Internal(e.to_string()))?;
-
-        // surface user-triggerable reverts distinctly from internal node faults
-        match result.result {
-            ExecutionResult::Success { output, .. } => Ok(output.into_data()),
-            ExecutionResult::Revert { output, .. } => Err(EvmReadError::Revert {
-                reason: alloy::sol_types::decode_revert_reason(&output),
-                output,
-            }),
-            ExecutionResult::Halt { reason, gas_used } => Err(EvmReadError::Internal(format!(
-                "contract call halted: {reason:?} (gas {gas_used})"
-            ))),
-        }
-    }
-
-    /// Extract the epoch number from a header's nonce.
-    pub fn extract_epoch_from_header(header: &ExecHeader) -> Epoch {
-        let nonce: u64 = header.nonce.into();
-        (nonce >> 32) as u32
-    }
-
-    /// Read the curret epoch number from the [ConsensusRegistry] on-chain.
-    fn get_current_epoch_number<DB>(&self, evm: &mut TNEvm<DB>) -> EvmReadResult<u32>
-    where
-        DB: alloy_evm::Database,
-    {
-        let calldata = ConsensusRegistry::getCurrentEpochCall {}.abi_encode().into();
-        self.call_consensus_registry::<_, u32>(evm, calldata)
-    }
-
-    /// Read the curret epoch info from the [ConsensusRegistry] on-chain.
-    fn get_current_epoch_info<DB>(
-        &self,
-        evm: &mut TNEvm<DB>,
-    ) -> EvmReadResult<ConsensusRegistry::EpochInfo>
-    where
-        DB: alloy_evm::Database,
-    {
-        let calldata = ConsensusRegistry::getCurrentEpochInfoCall {}.abi_encode().into();
-        self.call_consensus_registry::<_, ConsensusRegistry::EpochInfo>(evm, calldata)
-    }
-
-    /// Retrieve all `ValidatorInfo` in the committee for the provided epoch.
-    fn get_committee_validators_by_epoch<DB>(
-        &self,
-        epoch: Epoch,
-        evm: &mut TNEvm<DB>,
-    ) -> EvmReadResult<Vec<ConsensusRegistry::ValidatorInfo>>
-    where
-        DB: alloy_evm::Database,
-    {
-        let calldata = ConsensusRegistry::getCommitteeValidatorsCall { epoch }.abi_encode().into();
-        self.call_consensus_registry::<_, Vec<ConsensusRegistry::ValidatorInfo>>(evm, calldata)
-    }
-
-    /// Retrieve BLS pubkeys for the committee of the provided epoch.
-    fn get_committee_bls_pubkeys_by_epoch<DB>(
-        &self,
-        epoch: Epoch,
-        evm: &mut TNEvm<DB>,
-    ) -> EvmReadResult<Vec<alloy::primitives::Bytes>>
-    where
-        DB: alloy_evm::Database,
-    {
-        let calldata = ConsensusRegistry::getCommitteeBlsPubkeysCall { epoch }.abi_encode().into();
-        self.call_consensus_registry::<_, Vec<alloy::primitives::Bytes>>(evm, calldata)
-    }
-
-    /// Read fee configs for all workers from the [`WorkerConfigs`] contract at the given header.
-    ///
-    /// Builds an EVM against `header`'s state and issues a single `getAllWorkerConfigs()` call.
-    /// Returns the on-chain worker count alongside the decoded [`WorkerFeeConfig`]s.
-    ///
-    /// Failures are classified per [`StateReadError`]. The classification boundary: the state
-    /// provider construction and any database fault the EVM hits while lazily reading state are
-    /// [`StateReadError::Provider`] (node-local — peers reading the same block may succeed);
-    /// everything downstream of a successfully-executing EVM — contract absent (empty return
-    /// data), revert, halt, ABI decode, arity mismatch — plus EVM environment construction is
-    /// [`StateReadError::ChainGlobal`] (a deterministic product of the pinned block, identical on
-    /// every node).
-    fn worker_fee_configs_inner(
-        &self,
-        header: &SealedHeader,
-    ) -> StateReadResult<(usize, Vec<WorkerFeeConfig>)> {
-        let state_provider =
-            self.inner.blockchain_provider.state_by_block_hash(header.hash()).map_err(|e| {
-                StateReadError::Provider(format!("state provider at {}: {e}", header.hash()))
-            })?;
-        let state = StateProviderDatabase::new(&state_provider);
-        let mut db = State::builder().with_database(state).with_bundle_update().build();
-        let evm_env = self.inner.evm_config.evm_env(header).map_err(|e| {
-            StateReadError::ChainGlobal(format!("evm env for {}: {e}", header.hash()))
-        })?;
-        let mut tn_evm = self.inner.evm_config.evm_factory().create_evm(&mut db, evm_env);
-
-        let calldata = WorkerConfigs::getAllWorkerConfigsCall {}.abi_encode().into();
-        let result = Self::classified_system_call(
-            &mut tn_evm,
-            SYSTEM_ADDRESS,
-            WORKER_CONFIGS_ADDRESS,
-            calldata,
-        )?;
-        let data = match result.result {
-            ExecutionResult::Success { output, .. } => output.into_data(),
-            e => {
-                return Err(StateReadError::ChainGlobal(format!(
-                    "failed to read worker configs: {e:?}"
-                )))
-            }
-        };
-        let ret =
-            <WorkerConfigs::getAllWorkerConfigsCall as alloy::sol_types::SolCall>::abi_decode_returns(
-                &data,
-            )
-            .map_err(|e| {
-                StateReadError::ChainGlobal(format!(
-                    "worker configs return decode failed (contract absent at this block?): {e}"
-                ))
-            })?;
-
-        let num_workers = ret.count as usize;
-        if ret.strategies.len() != num_workers
-            || ret.values.len() != num_workers
-            || ret.datas.len() != num_workers
-        {
-            return Err(StateReadError::ChainGlobal(format!(
-                "worker config arity mismatch: count={num_workers}, strategies={}, values={}, datas={}",
-                ret.strategies.len(),
-                ret.values.len(),
-                ret.datas.len(),
-            )));
-        }
-
-        let mut configs = Vec::with_capacity(num_workers);
-        for (worker_id, (&strategy, &value)) in
-            ret.strategies.iter().zip(ret.values.iter()).enumerate()
-        {
-            let config = match strategy {
-                0 => WorkerFeeConfig::Eip1559 { target_gas: value },
-                1 => WorkerFeeConfig::Static { fee: value },
-                s => {
-                    // The contract rejects unknown strategies, so this branch only fires when a
-                    // future contract version introduces a strategy this node hasn't been
-                    // updated to understand. Fall back to EIP-1559 to preserve liveness instead
-                    // of halting all validators.
-                    tracing::warn!(
-                        target: "tn::reth",
-                        worker_id,
-                        strategy = s,
-                        "unknown fee strategy; falling back to strategy 0 (Eip1559)"
-                    );
-                    WorkerFeeConfig::Eip1559 { target_gas: value }
-                }
-            };
-            configs.push(config);
-        }
-
-        Ok((num_workers, configs))
-    }
-
-    /// Read worker fee configs from the [`WorkerConfigs`] contract at the block identified by
-    /// `block_hash`.
-    ///
-    /// Returns the on-chain worker count and one [`WorkerFeeConfig`] per worker. Failures are
-    /// classified per [`StateReadError`] (see [`Self::worker_fee_configs_inner`]); `block_hash`
-    /// failing to resolve to a sealed header is [`StateReadError::Provider`] — the pinned block
-    /// exists on the committee by construction, so a miss reflects this node's local view, not a
-    /// chain-global fact.
-    pub fn get_worker_fee_configs_at_block(
-        &self,
-        block_hash: B256,
-    ) -> StateReadResult<(usize, Vec<WorkerFeeConfig>)> {
-        let header = self
-            .sealed_header_by_hash(block_hash)
-            .map_err(|e| {
-                StateReadError::Provider(format!("header lookup for {block_hash:?}: {e}"))
-            })?
-            .ok_or_else(|| {
-                StateReadError::Provider(format!(
-                    "sealed header not found for block hash {block_hash:?}"
-                ))
-            })?;
-        self.worker_fee_configs_inner(&header)
-    }
-
-    /// Read the [`ConsensusRegistry`] [`EpochInfo`](ConsensusRegistry::EpochInfo) for `epoch` at
-    /// the block identified by `block_hash`.
-    ///
-    /// Builds an EVM pinned to `block_hash`'s state and issues a single `getEpochInfo(uint32)`
-    /// call. The registry keeps a ring buffer of the four most recent epochs and reverts
-    /// (`InvalidEpoch`) for anything outside it, so a successful return is guaranteed to be the
-    /// requested epoch's record. Used by the epoch manager to recover the previous epoch's block
-    /// range from its closing block when deriving next-epoch base fees.
-    ///
-    /// Fails with a descriptive error if `block_hash` does not resolve to a sealed header or the
-    /// registry call does not succeed.
-    pub fn get_epoch_info_at_block(
-        &self,
-        epoch: Epoch,
-        block_hash: B256,
-    ) -> eyre::Result<ConsensusRegistry::EpochInfo> {
-        let header = self
-            .sealed_header_by_hash(block_hash)?
-            .ok_or_else(|| eyre::eyre!("sealed header not found for block hash {block_hash:?}"))?;
-        let state_provider = self.inner.blockchain_provider.state_by_block_hash(header.hash())?;
-        let state = StateProviderDatabase::new(&state_provider);
-        let mut db = State::builder().with_database(state).with_bundle_update().build();
-        let mut tn_evm = self
-            .inner
-            .evm_config
-            .evm_factory()
-            .create_evm(&mut db, self.inner.evm_config.evm_env(&header)?);
-
-        let calldata = ConsensusRegistry::getEpochInfoCall { epoch }.abi_encode().into();
-        self.call_consensus_registry::<_, ConsensusRegistry::EpochInfo>(&mut tn_evm, calldata)
-            .map_err(Into::into)
-    }
-
-    /// Read worker fee configs from the [`WorkerConfigs`] contract at the canonical tip.
-    ///
-    /// The returned `Vec`'s length is the on-chain `numWorkers()` at the canonical tip (the
-    /// arity between the count and the per-worker arrays is validated in
-    /// [`Self::worker_fee_configs_inner`]). Callers size their in-memory worker state (e.g. the
-    /// `GasAccumulator`) to match, rather than asserting a preconceived count.
-    pub fn get_worker_fee_configs(&self) -> StateReadResult<Vec<WorkerFeeConfig>> {
-        let canonical_tip = self.canonical_tip();
-        let (_num_workers, configs) = self.worker_fee_configs_inner(&canonical_tip)?;
-        Ok(configs)
-    }
-
-    /// Read the CURRENT epoch number and [`EpochInfo`](ConsensusRegistry::EpochInfo) from the
-    /// [`ConsensusRegistry`] at `header`, with failures classified per [`StateReadError`].
-    ///
-    /// This is the close-time identity read for the epoch manager's `adjust_base_fees`: at an
-    /// epoch's closing block the registry state has already crossed to the entered epoch
-    /// (`concludeEpoch` ran inside that block), so the returned info is the entered epoch's record
-    /// and its `blockHeight` must equal `header.number + 1`. It reads exactly what
-    /// [`Self::epoch_state_at_header`] reads for the same check but skips the committee/BLS/
-    /// epoch-start lookups (a gating check needs only the epoch identity) and — unlike that
-    /// method, whose failures collapse into `eyre` strings — keeps node-local provider faults
-    /// (NOT committee-deterministic: retry or halt) distinguishable from chain-global failures
-    /// (committee-deterministic: fail-open stays consistent).
-    pub fn get_current_epoch_info_at_header(
-        &self,
-        header: &SealedHeader,
-    ) -> StateReadResult<(Epoch, ConsensusRegistry::EpochInfo)> {
-        let state_provider =
-            self.inner.blockchain_provider.state_by_block_hash(header.hash()).map_err(|e| {
-                StateReadError::Provider(format!("state provider at {}: {e}", header.hash()))
-            })?;
-        let state = StateProviderDatabase::new(&state_provider);
-        let mut db = State::builder().with_database(state).with_bundle_update().build();
-        let evm_env = self.inner.evm_config.evm_env(header).map_err(|e| {
-            StateReadError::ChainGlobal(format!("evm env for {}: {e}", header.hash()))
-        })?;
-        let mut tn_evm = self.inner.evm_config.evm_factory().create_evm(&mut db, evm_env);
-
-        // both reads observe the ONE pinned EVM state
-        let epoch: Epoch = Self::classified_registry_read(
-            &mut tn_evm,
-            ConsensusRegistry::getCurrentEpochCall {}.abi_encode().into(),
-        )?;
-        let epoch_info: ConsensusRegistry::EpochInfo = Self::classified_registry_read(
-            &mut tn_evm,
-            ConsensusRegistry::getCurrentEpochInfoCall {}.abi_encode().into(),
-        )?;
-        Ok((epoch, epoch_info))
-    }
-
-    /// Execute a read-only [`ConsensusRegistry`] call on `evm` and decode the result, classifying
-    /// failures per [`StateReadError`].
-    ///
-    /// The [`StateReadError`]-typed sibling of [`Self::call_consensus_registry`]: revert, halt,
-    /// and decode failures are all deterministic products of the pinned chain state
-    /// ([`StateReadError::ChainGlobal`]); only a database fault inside the EVM (via
-    /// [`Self::classified_system_call`]) is node-local ([`StateReadError::Provider`]).
-    fn classified_registry_read<DB, T>(evm: &mut TNEvm<DB>, calldata: Bytes) -> StateReadResult<T>
-    where
-        DB: alloy_evm::Database,
-        T: alloy::sol_types::SolValue,
-        T: From<
-            <<T as alloy::sol_types::SolValue>::SolType as alloy::sol_types::SolType>::RustType,
-        >,
-    {
-        let state = Self::classified_system_call(
-            evm,
-            SYSTEM_ADDRESS,
-            CONSENSUS_REGISTRY_ADDRESS,
-            calldata,
-        )?;
-        match state.result {
-            ExecutionResult::Success { output, .. } => {
-                alloy::sol_types::SolValue::abi_decode(&output.into_data()).map_err(|e| {
-                    StateReadError::ChainGlobal(format!(
-                        "registry return decode failed (contract absent at this block?): {e}"
-                    ))
-                })
-            }
-            ExecutionResult::Revert { output, .. } => Err(StateReadError::ChainGlobal(format!(
-                "registry call reverted: {:?}",
-                alloy::sol_types::decode_revert_reason(&output)
-            ))),
-            ExecutionResult::Halt { reason, gas_used } => Err(StateReadError::ChainGlobal(
-                format!("registry call halted: {reason:?} (gas {gas_used})"),
-            )),
-        }
-    }
-
-    /// Execute a read-only system call on `evm`, classifying failures per [`StateReadError`].
-    ///
-    /// An [`EVMError::Database`] is a node-local provider fault surfaced by the EVM's lazy state
-    /// reads (the state provider is only CONSTRUCTED up front; account/storage/bytecode loads
-    /// happen during execution, so an MDBX/provider I/O fault lands here) — classified
-    /// [`StateReadError::Provider`]. Every other transact failure derives deterministically from
-    /// the pinned block and calldata, so it is [`StateReadError::ChainGlobal`].
-    fn classified_system_call<DB>(
-        evm: &mut TNEvm<DB>,
-        caller: Address,
-        contract: Address,
-        calldata: Bytes,
-    ) -> StateReadResult<ResultAndState>
-    where
-        DB: alloy_evm::Database,
-    {
-        evm.transact_system_call(caller, contract, calldata).map_err(|e| match e {
-            EVMError::Database(db_err) => {
-                StateReadError::Provider(format!("system call state read failed: {db_err}"))
-            }
-            other => {
-                StateReadError::ChainGlobal(format!("system call failed reading state: {other}"))
-            }
-        })
-    }
-
-    /// Helper function to call `ConsensusRegistry` state on-chain.
-    fn call_consensus_registry<DB, T>(
-        &self,
-        evm: &mut TNEvm<DB>,
-        calldata: Bytes,
-    ) -> EvmReadResult<T>
-    where
-        DB: alloy_evm::Database,
-        T: alloy::sol_types::SolValue,
-        T: From<
-            <<T as alloy::sol_types::SolValue>::SolType as alloy::sol_types::SolType>::RustType,
-        >,
-    {
-        let state = self
-            .read_state_on_chain(evm, SYSTEM_ADDRESS, CONSENSUS_REGISTRY_ADDRESS, calldata)
-            .map_err(|e| EvmReadError::Internal(e.to_string()))?;
-
-        // retrieve data from state, distinguishing user-triggerable reverts from node faults
-        match state.result {
-            ExecutionResult::Success { output, .. } => {
-                let data = output.into_data();
-                // use SolValue to decode the result
-                alloy::sol_types::SolValue::abi_decode(&data).map_err(|e| {
-                    EvmReadError::Internal(format!("registry return decode failed: {e}"))
-                })
-            }
-            ExecutionResult::Revert { output, .. } => Err(EvmReadError::Revert {
-                reason: alloy::sol_types::decode_revert_reason(&output),
-                output,
-            }),
-            ExecutionResult::Halt { reason, gas_used } => Err(EvmReadError::Internal(format!(
-                "registry call halted: {reason:?} (gas {gas_used})"
-            ))),
-        }
-    }
-
-    /// Read state on-chain.
-    fn read_state_on_chain<DB>(
-        &self,
-        evm: &mut TNEvm<DB>,
-        caller: Address,
-        contract: Address,
-        calldata: Bytes,
-    ) -> TnRethResult<ResultAndState>
-    where
-        DB: alloy_evm::Database,
-    {
-        // read from state
-        let res = match evm.transact_system_call(caller, contract, calldata) {
-            Ok(res) => res,
-            Err(e) => {
-                // fatal error
-                error!(target: "engine", ?caller, ?contract, "failed to read state: {}", e);
-                return Err(TnRethError::EVMCustom(format!(
-                    "system call failed reading state: {e}"
-                )));
-            }
-        };
-
-        Ok(res)
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use std::collections::VecDeque;
 
     use super::*;
     use crate::{
+        payload::TNPayload,
         system_calls::ConsensusRegistry::ValidatorStatus,
         test_utils::{
             execute_payload_and_update_canonical_chain, governance_burn_tx,
@@ -2936,10 +1009,9 @@ mod tests {
             .with_bundle_update()
             .build();
         let mut tn_evm = reth_env
-            .inner
-            .evm_config
+            .evm_config()
             .evm_factory()
-            .create_evm(&mut db, reth_env.inner.evm_config.evm_env(canonical_header.header())?);
+            .create_evm(&mut db, reth_env.evm_config().evm_env(canonical_header.header())?);
 
         // read new committee (always 2 epochs ahead)
         let calldata = ConsensusRegistry::getEpochInfoCall { epoch: epoch + 1 }.abi_encode().into();
@@ -3016,10 +1088,9 @@ mod tests {
             .with_bundle_update()
             .build();
         let mut tn_evm = reth_env
-            .inner
-            .evm_config
+            .evm_config()
             .evm_factory()
-            .create_evm(&mut db, reth_env.inner.evm_config.evm_env(canonical_header.header())?);
+            .create_evm(&mut db, reth_env.evm_config().evm_env(canonical_header.header())?);
 
         // assert validator 2 is pending exit
         let calldata =
@@ -3083,10 +1154,9 @@ mod tests {
             .with_bundle_update()
             .build();
         let mut tn_evm = reth_env
-            .inner
-            .evm_config
+            .evm_config()
             .evm_factory()
-            .create_evm(&mut db, reth_env.inner.evm_config.evm_env(canonical_header.header())?);
+            .create_evm(&mut db, reth_env.evm_config().evm_env(canonical_header.header())?);
 
         // assert validator 2 is pending exit
         let calldata =
